refactor(tokens): removing tokens/row
