+++ packages/components/col/src/col.vue
@@ -8,7 +8,7 @@
 import { computed, inject } from 'vue'
 import { isNumber, isObject } from '@element-plus/utils'
 import { useNamespace } from '@element-plus/hooks'
-import { rowContextKey } from '@element-plus/tokens'
+import { rowContextKey } from '@element-plus/components/row'
 import { colProps } from './col'
 import type { CSSProperties } from 'vue'
