+++ packages/components/row/index.ts
@@ -5,3 +5,4 @@ export const ElRow = withInstall(Row)
 export default ElRow
 
 export * from './src/row'
+export * from './src/constants'
