+++ packages/components/row/src/row.vue
@@ -7,7 +7,7 @@
 <script setup lang="ts">
 import { computed, provide } from 'vue'
 import { useNamespace } from '@element-plus/hooks'
-import { rowContextKey } from '@element-plus/tokens'
+import { rowContextKey } from './constants'
 import { rowProps } from './row'
 import type { CSSProperties } from 'vue'
