+++ packages/tokens/index.ts
@@ -1,4 +1,3 @@
-export * from './row'
 export * from './scrollbar'
 export * from './slider'
 export * from './tabs'
